% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{addGenome}
\alias{addGenome}
\title{Add your own Genome}
\usage{
addGenome(genome, name = "userDefined")
}
\arguments{
\item{genome}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'userDefined'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Add your own Genome for infercna to use. The
}
\details{
DETAILS
}
