% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{as_genomic}
\alias{as_genomic}
\title{FUNCTION_TITLE}
\usage{
as_genomic(x = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
