% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{availableGenomes}
\alias{availableGenomes}
\title{List of available genomes}
\usage{
availableGenomes()
}
\value{
character vector
}
\description{
See which genomes are available for use
}
\details{
more genomes will be added in future dev patches.
}
