% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt771.R
\docType{data}
\name{bt771}
\alias{bt771}
\title{scRNA-seq data from a patient with Glioblastoma}
\format{A matrix with 8556 genes (rows) and 1265 cells (columns):}
\source{
\url{https://portals.broadinstitute.org/single_cell/study/SCP393/single-cell-rna-seq-of-adult-and-pediatric-glioblastoma}
}
\usage{
bt771
}
\description{
A single-cell RNA-sequencing dataset containing 8556 HQ genes and 1265 high quality cells. The data was generated using the Smart-Seq2 protocol and was analysed in the form log2(TPM/10 + 1).
}
\keyword{datasets}
