% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{clip}
\alias{clip}
\title{FUNCTION_TITLE}
\usage{
clip(mat, range = c(-3, 3))
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{range}{PARAM_DESCRIPTION, Default: c(-3, 3)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
