% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{clones}
\alias{clones}
\title{FUNCTION_TITLE}
\usage{
clones(m = NULL, L = NULL, by = "arm", prob = 0.95,
  coverage = 0.8, size = 10, boolean = FALSE)
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{L}{PARAM_DESCRIPTION, Default: NULL}

\item{by}{PARAM_DESCRIPTION, Default: 'arm'}

\item{prob}{PARAM_DESCRIPTION, Default: 0.95}

\item{coverage}{PARAM_DESCRIPTION, Default: 0.8}

\item{size}{PARAM_DESCRIPTION, Default: 10}

\item{boolean}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
