% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMalignant_by_score.R
\name{cna_classify_cells}
\alias{cna_classify_cells}
\title{Classify CNA scores}
\usage{
cna_classify_cells(
  cna_scores,
  signal_threshold = 0.05,
  correlation_threshold = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{cna_scores}{cna scores info get from \code{\link[Infercna]{cna_compute_scores}}}

\item{signal_threshold}{The threshold of cna scores}

\item{correlation_threshold}{The threshold of cna correlation}

\item{verbose}{print progress messages. Default: FALSE}
}
\value{
return A \code{tibble} with five columns:
\enumerate{
  \item CellID
  \item Signal
  \item Correlation
  \item CNADetected
  \item Malignant
}
}
\description{
This function classifies the CNA scores according to the supplied signal and correlation thresholds.
}
\seealso{
\link{cna_compute_scores}
}
