% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMalignant_by_score.R
\name{cna_classify_cells_with_cluster}
\alias{cna_classify_cells_with_cluster}
\title{Identify Malignant cells with cluster info}
\usage{
cna_classify_cells_with_cluster(
  cna_scores,
  clusters = NULL,
  signal_threshold = 0.05,
  correlation_threshold = 0.05,
  min_cluster_cc_freq = 0.5,
  cna_matrix = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cna_scores}{cna scores info get from \code{\link[Infercna]{cna_compute_scores}} or \code{\link[Infercna]{cna_classify_cells}}}

\item{clusters}{a named vector of cell clusters with cell id, this must be supplied}

\item{signal_threshold}{The threshold of cna scores, this is used when \code{cna_scores} have no malignant info}

\item{correlation_threshold}{The threshold of cna correlation, this is used when \code{cna_scores} have no malignant info}

\item{min_cluster_cc_freq}{The minium cancer cell or normal cell precent in a cluster, Default 0.5}

\item{cna_matrix}{The copy number matrix}

\item{verbose}{print progress messages. Default: FALSE}
}
\value{
return A \code{tibble} with \code{Malignant} column show the cell type info
}
\description{
Identify Malignant cells by cna signal and cna correlation with cluster info
}
\seealso{
\code{\link[Infercna]{cna_compute_scores}} \code{\link[Infercna]{cna_classify_cells}}
}
