% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMalignant_by_score.R
\name{cna_scores_plot}
\alias{cna_scores_plot}
\title{Visualise of cna signal and correlation}
\usage{
cna_scores_plot(
  cna_scores,
  signal_threshold = 0.05,
  correlation_threshold = 0.5
)
}
\arguments{
\item{cna_scores}{cna scores info get from \code{\link[Infercna]{cna_compute_scores}} or \code{\link[Infercna]{cna_classify_cells}}}

\item{signal_threshold}{The threshold of cna scores}

\item{correlation_threshold}{The threshold of cna correlation}
}
\value{
return a ggplot object
}
\description{
Visualise Malignant and Non-Malignant Subsets of cells. This is achieved by plotting, for each cell, its CNA signal over its CNA correlation. Please see \link{cna_compute_scores}.
}
\seealso{
\code{\link[Infercna]{cna_compute_scores}} \code{\link[Infercna]{cna_classify_cells}}
}
