% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaHotspotCells.R
\name{cnaHotspotCells}
\alias{cnaHotspotCells}
\title{Find the Cells with the highest CNA signal}
\usage{
cnaHotspotCells(cna, cell.quantile, gene.quantile = NULL)
}
\arguments{
\item{cna}{a matrix of cell rows by cell columns containing CNA values}

\item{cell.quantile}{calculate CNA measures including only top / "hotspot" cells according to their squared CNA values across all genes. Value between 0 and 1 denoting the quantile of cells to include.}
}
\value{
cell names in the top nth quantile, where n is specified via <cell.quantile>
}
\description{
Find Cells in the top nth quantile for CNA Signal values
}
