% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaHotspotGenes.R
\name{cnaHotspotGenes}
\alias{cnaHotspotGenes}
\title{FUNCTION_TITLE}
\usage{
cnaHotspotGenes(cna, threshold)
}
\arguments{
\item{cna}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
