% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaParamsPlot.R
\name{cnaParamsPlot}
\alias{cnaParamsPlot}
\title{Plot CNA parameters to evaluate cell CNAs}
\usage{
cnaParamsPlot(
  data,
  x,
  y,
  group = NULL,
  points = TRUE,
  point.col = "black",
  point.alpha = 0.075,
  point.size = 0.5,
  xlab = "CNA correlation",
  ylab = "CNA signal",
  xlim = NULL,
  ylim = NULL,
  xintercept = NULL,
  yintercept = NULL,
  bins = 30,
  breaks = scales::breaks_pretty(n = 4),
  xbreaks = breaks,
  ybreaks = ggplot2::waiver(),
  contour = TRUE,
  geom = "polygon",
  legend.justification = "right",
  legend.position = "top",
  legend.title = ggplot2::waiver(),
  cols = c("red", "blue"),
  text.size = 16,
  label.hjust = 1.25,
  label.vjust = 1.25,
  label.size = 5,
  ...
)
}
\arguments{
\item{data}{data.frame with x (CNA correlations), y (CNA signals), group (e.g. presumed Malignant versus Reference Normal) and (optional) sample columns}

\item{x}{column name corresponding to x values. Either passed as a string or as an unquoted column name. eg 'x' or x is fine.}

\item{y}{column name corresponding to y values. Either passed as a string or as an unquoted column name. eg 'y' or y is fine.}

\item{group}{column name corresponding to groups. Either passed as a string or as an unquoted column name. eg 'group' or group is fine.}

\item{points}{logical; should cell points be plotted as well as the contour plot? Default: TRUE}

\item{point.col}{relevant only if points = TRUE. specify point colour. Default: 'black'}

\item{point.alpha}{relevant only if points = TRUE. specify point transparency. Default: 0.075}

\item{point.size}{relevant only if points = TRUE. specify point size. Default: 0.5}

\item{xlab}{x-axis title. Default: 'CNA correlation'}

\item{ylab}{y-axis title. Default: 'CNA signal'}

\item{xlim}{x-axis limits. Default: NULL}

\item{ylim}{y-axis limits. Default: NULL}

\item{xintercept}{PARAM_DESCRIPTION, Default: NULL}

\item{yintercept}{PARAM_DESCRIPTION, Default: NULL}

\item{bins}{number of bins to use for contour levels. Fewer bins averages over more of the data but often results with a 'cleaner' looking contour. Default: 7}

\item{breaks}{breaks. Default: scales::breaks_pretty(n = 4)}

\item{xbreaks}{x-axis breaks. Default: breaks}

\item{ybreaks}{y-axis breaks. Default: ggplot2::waiver()}

\item{contour}{contour lines as well as filled in contour areas? Default: TRUE}

\item{geom}{I've only used polygon in this specific function but you can go crazy and try other things. Default: 'polygon'}

\item{legend.justification}{legend justification. Default: 'right'}

\item{legend.position}{legend position. set to 'none' if not desired. Default: 'top'}

\item{legend.title}{legend title Default: ggplot2::waiver()}

\item{cols}{colours of each group. Note that if more than two groups are present, more colours should be provided than the default two. Default: c("red", "blue")}

\item{text.size}{base text size. Default: 16}

\item{label.hjust}{only relevant if sample column in data. Adjust position of geom_text with sample name. Default: 1.25}

\item{label.size}{only relevant if sample column in data. geom_text label size. Default: 5}

\item{...}{other arguments passed to ggplot2::facet_wrap(.~sample, ...) if sample column in data.}
}
\value{
ggplot2 object
}
\description{
Plot cells according to their CNA correlation and CNA signal values, and display this as a contour plot. If a "sample" column is provided, the resulting plot will be faceted by sample, with the sample name written in the plot. A "group" column is currently required (suggested to just make a group column with all values the same if there are no real groups annotated).
}
\seealso{
\code{\link[scales]{breaks_pretty}}
 \code{\link[ggplot2]{waiver}},\code{\link[ggplot2]{tidyeval}},\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_density_2d}},\code{\link[ggplot2]{geom_label}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{scale_alpha}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}},\code{\link[ggplot2]{guide_legend}},\code{\link[ggplot2]{facet_wrap}}
}
