% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaPlot.R
\name{cnaPlot}
\alias{cnaPlot}
\title{Plot a Heatmap of CNA Values}
\usage{
cnaPlot(cna, limits = c(-1, 1), ratio = 0.5, cols = heatCols,
  x.name = "Chromosome", y.name = "Cell",
  legend.title = "Inferred CNA\\n[log2 ratio]", x.hide = c("13", "18",
  "21", "Y"), orderCells = F, order.with = NULL, euclid.dist = F,
  angle = NULL, x.angle = NULL, y.angle = 0, axis.rel = 1,
  base.size = 12, axis.title.size = 12, axis.text.size = 11,
  base.col = "#073642", title = NULL, subtitle = NULL,
  caption = NULL, text.size = 12, y.hide = NULL, tile.size = 0.1,
  tile.col = NULL, legend.position = "right", legend.height = 2,
  legend.width = 0.6, legend.rel = 0.9, legend.colour = "black",
  legend.breaks = NULL, legend.labels = NULL,
  legend.justification = "top", legend.title.position = "bottom",
  legend.title.angle = NULL, legend.title.rel = 0.9)
}
\arguments{
\item{cna}{matrix of CNA values (genes by cells).}

\item{limits}{colour range. Cells >= upper limit will be the same colour, and likewise for cells <= lower limit. Default: c(-1, 1)}

\item{ratio}{aspect ratio of the panel. Default: 0.7}

\item{cols}{character vector of colours to use. Default: heatCols}

\item{x.name}{x axis label. Default: 'Chromosome'}

\item{y.name}{y axis label. Default: 'Cell'}

\item{legend.title}{PARAM_DESCRIPTION, Default: 'Inferred CNA `[log2 ratio`]'}

\item{x.hide}{PARAM_DESCRIPTION, Default: c("13", "18", "21", "Y")}

\item{orderCells}{boolean value indicating whether cells should be ordered by hierarchical clustering. Default: F}

\item{angle}{angle of axes tick labels. x.angle and y.angle inherit from angle. Default: NULL}

\item{x.angle}{angle of x axis tick labels. If left, will inherit from angle. Default: NULL}

\item{y.angle}{angle of y axis tick labels. If left, will inherit from angle. Default: 0}

\item{axis.rel}{PARAM_DESCRIPTION, Default: 1}

\item{base.size}{PARAM_DESCRIPTION, Default: 12}

\item{axis.title.size}{PARAM_DESCRIPTION, Default: 12}

\item{axis.text.size}{PARAM_DESCRIPTION, Default: 11}

\item{base.col}{PARAM_DESCRIPTION, Default: '#073642'}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{caption}{PARAM_DESCRIPTION, Default: NULL}

\item{text.size}{PARAM_DESCRIPTION, Default: 12}

\item{y.hide}{PARAM_DESCRIPTION, Default: NULL}

\item{tile.size}{PARAM_DESCRIPTION, Default: 0.1}

\item{tile.col}{PARAM_DESCRIPTION, Default: NULL}

\item{legend.position}{PARAM_DESCRIPTION, Default: 'right'}

\item{legend.height}{PARAM_DESCRIPTION, Default: 2}

\item{legend.width}{PARAM_DESCRIPTION, Default: 0.6}

\item{legend.rel}{PARAM_DESCRIPTION, Default: 0.9}

\item{legend.colour}{PARAM_DESCRIPTION, Default: 'black'}

\item{legend.breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{legend.labels}{PARAM_DESCRIPTION, Default: NULL}

\item{legend.justification}{PARAM_DESCRIPTION, Default: 'top'}

\item{legend.title.position}{PARAM_DESCRIPTION, Default: 'bottom'}

\item{legend.title.angle}{PARAM_DESCRIPTION, Default: NULL}

\item{legend.title.rel}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Plot a heatmap of CNA values.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scrabble]{clusta}}
 \code{\link[reshape2]{melt}}
 \code{\link[ggpubr]{rotate_axis_text}}
}
