% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaScatterPlot.R
\name{cnaScatterPlot}
\alias{cnaScatterPlot}
\title{Visualise Malignant and Non-Malignant Subsets}
\usage{
cnaScatterPlot(
  cna,
  cor.method = "pearson",
  threshold = NULL,
  cor.threshold = threshold,
  signal.threshold = threshold,
  group = NULL,
  group.col = scalop::discrete_colours[1:length(group)],
  cex = 0.8,
  pch = 20,
  hline = NULL,
  vline = NULL,
  refCells = NULL,
  samples = NULL,
  ...
)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}

\item{cor.method}{character string indicating the method to use for the pairwise correlations. E.g. 'pearson', 'spearman'. Default: 'pearson'}

\item{threshold}{Default: NULL}

\item{cor.threshold}{PARAM_DESCRIPTION, Default: threshold}

\item{signal.threshold}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{group.col}{PARAM_DESCRIPTION, Default: 'magenta'}

\item{hline}{PARAM_DESCRIPTION, Default: NULL}

\item{vline}{PARAM_DESCRIPTION, Default: NULL}

\item{refCells}{a character vector of cell ids to exclude from average CNA profile that each cell is correlated to. You can pass reference normal cell ids to this argument if these are known. Default: NULL}

\item{samples}{if CNA correlations should be calculated within cell subgroups, provide i) a list of cell id groups, ii) a character vector of sample names to group cells by, iii) TRUE to extract sample names from cell ids and subsequently group. Default: NULL}

\item{...}{other arguments passed to scalop::unique_sample_names if samples = TRUE.}
}
\value{
a base R plot. If return value is saved to a variable, instead returns data points for cna correlations and cna signal in list form.
}
\description{
Visualise Malignant and Non-Malignant Subsets of cells. This is achieved by plotting, for each cell, its CNA signal over its CNA correlation. Please see `infercna::cnaSignal` and `infercna::cnaCor` for details.
}
