% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaSignal.R
\name{cnaSignal}
\alias{cnaSignal}
\title{Calculate the Means of Squared CNA Values}
\usage{
cnaSignal(cna, gene.quantile = NULL, refCells = NULL, samples = NULL)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}

\item{gene.quantile}{calculate CNA measures including only top / "hotspot" genes according to their squared CNA values across all cells. Value between 0 and 1 denoting the quantile of genes to include. Default: NULL}

\item{refCells}{a character vector of cell ids (e.g. normal reference cell ids) to exclude from calculation of cnaHotspotGenes. Only relevant if gene.quantile is not NULL. Default: NULL}

\item{samples}{if cnaHotspotGenes should be calculated within cell subgroups, provide i) a list of cell id groups, ii) a character vector of sample names to group cells by, iii) TRUE to extract sample names from cell ids and subsequently group. Default: NULL}

\item{...}{other arguments passed to scalop::unique_sample_names if samples = TRUE.}
}
\value{
a numeric vector of CNA signal values or the Mean of Squared CNA values
}
\description{
Calculate the Mean of Squared CNA Values
}
