% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{colCenter}
\alias{colCenter}
\title{Center a matrix column-wise}
\usage{
colCenter(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
a matrix with colMeans == 0
}
\description{
Center a matrix column-wise
}
