% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{currentGenome}
\alias{currentGenome}
\title{List the current genome name}
\usage{
currentGenome()
}
\value{
string
}
\description{
E.g. "hg19" if the genome being used is hg19
}
\details{
Default genome is "hg19"
}
