% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{dipcells}
\alias{dipcells}
\title{Normal Cell IDs in a cohort of 28 Glioblastoma samples.}
\format{A list of length two containing cell IDs of two normal cell types.
\describe{
  \item{oligodendrocytes}{219 normal oligodendrocyte cells in the GBM cohort}
  \item{macrophages}{707 normal macrophage cells in the GBM cohort}
}}
\source{
\url{portals.broadinstitute.org/single_cell/study/SCP393/single-cell-rna-seq-of-adult-and-pediatric-glioblastoma}
}
\usage{
dipcells
}
\description{
A list of two character vectors containing cell IDs of the normal cells that were found in the GBM study.
}
\keyword{datasets}
