% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercna.R
\name{dipcenter}
\alias{dipcenter}
\title{FUNCTION_TITLE}
\usage{
dipcenter(v, Min, Max)
}
\arguments{
\item{v}{PARAM_DESCRIPTION}

\item{Min}{PARAM_DESCRIPTION}

\item{Max}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
