% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercna.R
\name{dipcorrect}
\alias{dipcorrect}
\title{FUNCTION_TITLE}
\usage{
dipcorrect(cna, ...)
}
\arguments{
\item{cna}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Correction with diploid cells' <dipcells> CNAs: the boundaries of dipcells CNA values are the boundaries for what should be considered a CNA of 0. Thus, if the boundary is -0.1 and 0.1, then a cell with CNA = -0.5 will be corrected to -0.4 and a cell with CNA value of 1 will be corrected to 0.9.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
