% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infercna.R
\name{.mostExpressedGenes}
\alias{.mostExpressedGenes}
\title{Get most expression gene}
\usage{
.mostExpressedGenes(m, ngenes)
}
\arguments{
\item{m}{a matrix of genes X cells containing scRNA-seq expression data}

\item{ngenes}{The number or the precent of gene to select}
}
\value{
return a vector of gene name with high expression
}
\description{
select the top n number of genes for infer cna analysis
}
