% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clones.R
\name{envglob}
\alias{envglob}
\title{FUNCTION_TITLE}
\usage{
envglob(genes = NULL)
}
\arguments{
\item{genes}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
