% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clones.R
\name{envmatch}
\alias{envmatch}
\title{FUNCTION_TITLE}
\usage{
envmatch(genes)
}
\arguments{
\item{genes}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
