% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{expandToClones}
\alias{expandToClones}
\title{expand clones}
\usage{
expandToClones(modes, greaterThan = NULL)
}
\arguments{
\item{modes}{The result of \code{\link[Infercna]{fetchModes}}}

\item{greaterThan}{The clone number, Default NULL}
}
\description{
Order the clones
}
