% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenes.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter Genes by their Genome Features}
\usage{
filterGenes(x, value, attribute, throw = F, strict = F)
}
\arguments{
\item{x}{a character vector of gene names or a matrix with gene row names.}

\item{value}{a value or a list of values that are the genome features to filter}

\item{attribute}{the name(s) of the genome feature(s) in question. One or more of of 'chr', 'arm', 'start', 'end'.}

\item{throw}{boolean value indicating whether the filtered genes should be thrown away rather than kept. Default: F}

\item{strict}{boolean value indicating whether the filtered genes should match all criteria instead of the default to match any criteria. Default: F}
}
\value{
filtered vector or a matrix with filtered rows, depending on class of <x>.
}
\description{
Filter genes to keep (the default) or to throw according to genomic features. For example, filter genes that are either on chromosome 7 or chromosome arm 2p.
}
