% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{findClones}
\alias{findClones}
\title{Find Genetic Clones}
\usage{
findClones(m, prob = 0.95, rate = 0.8, size = 10)
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{prob}{PARAM_DESCRIPTION, Default: 0.95}

\item{size}{PARAM_DESCRIPTION, Default: 10}

\item{L}{PARAM_DESCRIPTION, Default: NULL}

\item{by}{PARAM_DESCRIPTION, Default: 'arm'}

\item{coverage}{PARAM_DESCRIPTION, Default: 0.8}

\item{boolean}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
