% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{genmatch}
\alias{genmatch}
\title{FUNCTION_TITLE}
\usage{
genmatch(genes)
}
\arguments{
\item{genes}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
