% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clones.R
\name{gensplit}
\alias{gensplit}
\title{FUNCTION_TITLE}
\usage{
gensplit(m = NULL, genes = NULL, by = "arm")
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{genes}{PARAM_DESCRIPTION, Default: NULL}

\item{by}{PARAM_DESCRIPTION, Default: 'arm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
