% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaplot.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Plot A Heatmap with ggplot2}
\usage{
ggheatmap(
  dat,
  x = Var1,
  y = Var2,
  fill = value,
  limits = c(-0.5, 0.5),
  raster.labels = FALSE,
  raster.label.col = "black",
  lim.find = F,
  lim.sym = T,
  title = waiver(),
  subtitle = waiver(),
  caption = waiver(),
  x.name = NULL,
  y.name = NULL,
  x.angle = 45,
  axis.rel = 1,
  title.rel = 1.1,
  text.size = 12,
  aspect.ratio = NULL,
  cols = NULL,
  na.value = "white",
  plot.margin = margin(0, 0.15, 0, 0, "cm"),
  legend.position = "bottom",
  legend.justification = "right",
  legend.margin = margin(0, 0.15, 0, 0, "cm"),
  legend.title = NULL,
  legend.title.position = "top",
  legend.title.rel = 0.8,
  legend.key.height = grid::unit(0.4, "cm"),
  legend.key.width = grid::unit(0.6, "cm"),
  legend.text.rel = 0.8,
  legend.text.color = "black",
  legend.breaks = NULL,
  legend.labels = NULL,
  ticks.linewidth = 0.5,
  x.axis.position = "bottom",
  y.axis.position = "left",
  breaks = waiver(),
  x.breaks = waiver(),
  y.breaks = waiver(),
  labels = waiver(),
  x.labels = waiver(),
  y.labels = waiver(),
  num = T,
  y.num = num,
  x.num = num,
  expand = c(0, 0)
)
}
\arguments{
\item{dat}{A \code{data.frame} or a \code{tibble} to plot ggplot.}

\item{x}{The column name of variable mapping to x axis.}

\item{y}{The column name of variable mapping to y axis.}

\item{fill}{The column name of variable mapping to heatmap fill.}

\item{limits}{A vector to set the max and min value to plot heatmap.}

\item{raster.labels}{Whether label the value in heatmap. Defualt: 45.}

\item{raster.label.col}{The color of raster.labels in heatmap. Default: black.}

\item{lim.find}{If you want to auto set limit, set this argument as TRUE.}

\item{lim.sym}{Set with \code{lim.find}, whether use the same upper or lower threshold}

\item{title}{Heatmap title.}

\item{subtitle}{Heatmap subtitle.}

\item{caption}{Heatmap caption title.}

\item{x.name}{The x axis name.}

\item{y.name}{The y axis name.}

\item{x.angle}{The angle to change of x axis text. Default: TRUE}

\item{axis.rel}{The relative size of axis name. Default: 1}

\item{title.rel}{The relative size of title name. Default: 1.1}

\item{text.size}{base font size, given in pts.}

\item{aspect.ratio}{aspect ratio of the panel}

\item{cols}{a vector of color to use in heatmap}

\item{na.value}{Colour to use for missing values}

\item{plot.margin}{margin around entire plot \code{unit} with the sizes of the `top`, `right`, `bottom`, and `left` margins}

\item{legend.position}{the default position of legends `none`, `left`, `right`, `bottom`, `top`, `inside`}

\item{legend.justification}{anchor point for positioning legend inside plot `center` or two-element numeric vector or the justification according to the plot area when positioned outside the plot}

\item{legend.margin}{the margin around each legend \code{\link[ggplot2]{element}}}

\item{legend.title}{The legend title name}

\item{legend.title.position}{placement of legend title relative to the main legend. `top`, `right`, `bottom` or `left`.}

\item{legend.title.rel}{The relative size of legend title. Default: 0.8}

\item{legend.key.height, legend.key.width}{size of legend keys \code{unit}}

\item{legend.text.rel}{The relative size of legend text. Default: 0.8}

\item{legend.text.color}{The color of legend text. Default: black}

\item{legend.breaks}{The breaks of legend. One of:

- NULL for no breaks

- \code{waiver()} for the default breaks computed by the \code{\link[scales]{new_transform}}

- A numeric vector of positions

- A function that takes the limits as input and returns breaks as output e.g., a function returned by \code{\link[scales]{extended_breaks}}. Note that for position scales, limits are provided after scale expansion. Also accepts rlang \code{\link[rlang]{as_function}} function notation.

 @param legend.labels The labels of legend. One of:

- NULL for no labels

- \code{waiver()} for the default labels computed by the transformation object

- A character vector giving labels, must be same length as \code{breaks}

- An expression vector, must be the same length as breaks. See ?plotmath for details.

- A function that takes the breaks as input and returns labels as output. Also accepts rlang \code{\link[rlang]{as_function}} function notation.}

\item{ticks.linewidth}{The ticks width. Default: 0.5}

\item{x.axis.position, y.axis.position}{For position scales, The position of the axis. `top` or `bottom` for x axes, \code{left} or \code{right} for y axes.}

\item{breaks, x.breaks, y.breaks}{The breaks of x axis or y axis. If you set \code{breaks}, it will be use in \code{x.breaks} and \code{y.breaks}. Or you set them separately. One of:

- NULL for no breaks.

- \code{waiver()} for the default breaks computed by the \code{\link[scales]{new_transform}}

- A numeric vector of positions

- A function that takes the limits as input and returns breaks as output e.g., a function returned by \code{\link[scales]{extended_breaks}}. Note that for position scales, limits are provided after scale expansion. Also accepts rlang \code{\link[rlang]{as_function}} function notation.}

\item{labels, x.labels, y.labels}{The labels of x axis or y axis. If you set \code{breaks}, it will be use in \code{x.breaks} and \code{y.breaks}. Or you set them separately. One of:

- NULL for no breaks

- \code{waiver()} for the default labels computed by the \code{\link[scales]{new_transform}}

- A character vector giving labels, must be same length as \code{breaks}

- An expression vector, must be the same length as breaks. See ?plotmath for details.

- A function that takes the breaks as input and returns labels as output. Also accepts rlang \code{\link[rlang]{as_function}} function notation.}

\item{num, x.num, y.num}{Whether the set x axis or y axis as numeric, if you set \code{num}, it will be use in \code{x.num} and \code{y.num}. Or you set them separately. Default: TRUE}

\item{expand}{For position scales, a vector of range expansion constants used to add some padding around the data to ensure that they are placed some distance away from the axes.
Use the convenience function \code{expansion()} to generate the values for the expand argument. The defaults are to expand the scale by 5\% on each side for continuous variables,
and by 0.6 units on each side for discrete variables.}
}
\value{
return a ggplot object
}
\description{
`ggheatmap()` create a heatmap ggplot2 object.
}
