% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{hg19 reference genome gene position annotation.}
\format{
A data.frame with gene annotation information.
}
\source{
\url{https://grch37.ensembl.org}
}
\usage{
hg19
}
\description{
A data.frame with gene position inforamtion get from hg19 reference genome.
}
\keyword{datasets}
