% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{hg38 reference genome gene position annotation.}
\format{
A data.frame with gene annotation information.
}
\source{
\url{https://useast.ensembl.org}
}
\usage{
hg38
}
\description{
A data.frame with gene position inforamtion get from hg38 reference genome.
}
\keyword{datasets}
