% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercna.R
\name{infercna}
\alias{infercna}
\title{Infer Copy-Number Alterations From Single-Cell RNA-Seq Data}
\usage{
infercna(m, dipcells = NULL, window = 100, range = c(-3, 3))
}
\arguments{
\item{m}{a matrix of genes X cells RNA-seq expression data. The matrix should be row-centered.}

\item{dipcells}{a list of two or more character vectors, each containg cell IDs of normal cell types (one cell type per list element). Since these cell types are presumed diploid cells, their CNA values can be used to correct the remaining CNA values. Note that all cell IDs should correspond to column names in <m>. See infercna::dipcells (diploid cells) for an example. Default: NULL}

\item{window}{the size of the window to use for the rolling mean. Units are in number of genes. Default: 100}

\item{range}{values in <m> above and below this range will be set to values in range. Default: c(-3, 3)}
}
\value{
a matrix of genes X cells of inferred CNA values. Note that n = (window - 1)/2 genes will be lost from either extremity of the genome (ie. n genes lost at the start of chromosome 1 and at the end of chromosome Y, if the genome in question is H.sapiens.)
}
\description{
Infer Copy-Number Alterations From Single-Cell RNA-Seq Data
}
\details{
Correction with diploid cells' <dipcells> CNAs: the boundaries of dipcells CNA values are the boundaries for what should be considered a CNA of 0. Thus, if the boundary is -0.1 and 0.1, then a cell with CNA = -0.5 will be corrected to -0.4 and a cell with CNA value of 1 will be corrected to 0.9.
}
\examples{
\dontrun{
if(interactive()){
 dipcells = infercna::dipcells
 m = infercna::useData()
 cna = infercna::infercna(m = m, dipcells = dipcells)
 }
}
}
