% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMalignant.R
\name{isMalignant}
\alias{isMalignant}
\title{Find Malignant Subset of Cells}
\usage{
isMalignant(cna)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}
}
\description{
Find Malignant Subset of Cells according to their CNA parameters for correlation to Tumour of origin (using cnaCor()) and for the overall extent of CNA signal (the mean absolute CNA values; using cnaSignal()).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
