% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10}
\alias{mm10}
\title{mm10 reference genome gene position annotation.}
\format{
A data.frame with gene annotation information.
}
\source{
\url{https://feb2021.archive.ensembl.org}
}
\usage{
mm10
}
\description{
A data.frame with gene position inforamtion get from mm10 reference genome.
}
\keyword{datasets}
