% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm39}
\alias{mm39}
\title{mm39 reference genome gene position annotation.}
\format{
A data.frame with gene annotation information.
}
\source{
\url{https://useast.ensembl.org}
}
\usage{
mm39
}
\description{
A data.frame with gene position inforamtion get from mm39 reference genome.
}
\keyword{datasets}
