% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{colcenter}
\alias{rowcenter}
\alias{logtpm}
\alias{unlogtpm}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{scalop}{\code{\link[scalop]{colcenter}}, \code{\link[scalop]{logtpm}}, \code{\link[scalop]{rowcenter}}, \code{\link[scalop]{unlogtpm}}}
}}

