% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{retrieveGenome}
\alias{retrieveGenome}
\title{Retrieve genome data}
\usage{
retrieveGenome(name = NULL)
}
\arguments{
\item{name}{a genome name. Currently one of 'hg19' (human), 'hg38' (latest human), 'mm10' (mouse). Default: NULL}
}
\value{
a tibble
}
\description{
Returns a tibble dataframe of the current genome in use. If on function call a character string is supplied that corresponds to a genome in the package, that genome will instead be returned.
}
\seealso{
\code{\link[tibble]{as_tibble}}
}
