% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{runMean}
\alias{runMean}
\title{FUNCTION_TITLE}
\usage{
runMean(mat, k = 100, endrule = "mean", align = "center")
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION, Default: 100}

\item{endrule}{PARAM_DESCRIPTION, Default: 'mean'}

\item{align}{PARAM_DESCRIPTION, Default: 'center'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[caTools]{runmean}}
}
