% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlogtpm.R
\name{unlogtpm}
\alias{unlogtpm}
\title{Convert logTPM to TPM}
\usage{
unlogtpm(m, bulk = F)
}
\arguments{
\item{m}{matrix of logTPM values (gene rows; cell columns)}

\item{bulk}{if bulk then instead uses 2^(TPM)-1. i.e. no scaling. Default: F}
}
\value{
TPM matrix
}
\description{
Convert logTPM to TPM, i.e. using 10*(2^(TPM)-1).
}
\details{
TPM/10 is used for single cells since 100,000 is a more reasonable estimate than 1,000,000 for the number of RNA transcripts in a cell. 1,000,000 is reasonable estimate for bulk samples that contain multiple cells.
}
