% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{useGenome}
\alias{useGenome}
\title{Select a Genome for infercna to use}
\usage{
useGenome(name)
}
\arguments{
\item{name}{a character string of genome name. One of 'hg19', 'hg38', 'mm10'.}
}
\value{
genome variables are set internally. No return.
}
\description{
Select your genome of choice at the start of an analysis. The available genomes in the current implementation are, for H.sapiens, hg38 (latest) and hg19 (preceding) and for mouse, mm10. You can see which genomes are available via availableGenomes().
}
