% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGenome.R
\name{addGenome}
\alias{addGenome}
\title{Add your own Genome}
\usage{
addGenome(genome, name = "userDefined")
}
\arguments{
\item{genome}{genome data provided as a dataframe. The dataframe should contain columns 'symbol', 'start_position', 'end_position', 'chromosome_name', 'arm'. The columns 'chromosome_name' and 'arm' should be factors, with the chromosome/chromosome arms ordered correctly.}

\item{name}{a character string; the name of your genome. Default: 'userDefined'}
}
\value{
no return value. The genome variables will be updated internally.
}
\description{
Add your own Genome for infercna to use. The
}
