% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{clip}
\alias{clip}
\title{Squish matrix values into range}
\usage{
clip(m, range = c(-3, 3))
}
\arguments{
\item{m}{matrix to manipulate}

\item{range}{numeric vector of length two giving desired output range. Default: c(-3, 3)}
}
\description{
Squish matrix values into range
}
