% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaHotspotGenes.R
\name{cnaHotspotGenes}
\alias{cnaHotspotGenes}
\title{Find the Genes with the highest CNA signal}
\usage{
cnaHotspotGenes(cna, gene.quantile, cell.quantile = NULL)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values}

\item{gene.quantile}{calculate CNA measures including only top / "hotspot" genes according to their squared CNA values across all cells. Value between 0 and 1 denoting the quantile of genes to include.}
}
\value{
gene names in the top nth quantile, where n is specified via <gene.quantile>
}
\description{
Find Genes in the top nth quantile for CNA Signal values
}
