% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaPlot.R
\name{cnaPlot}
\alias{cnaPlot}
\title{Plot a Heatmap of CNA Values}
\usage{
cnaPlot(
  cna,
  limits = c(-1, 1),
  ratio = 0.5,
  cols = heatCols,
  x.name = "Chromosome",
  y.name = "Cell",
  legend.title = "Inferred CNA\\n[log2 ratio]",
  x.hide = c("13", "18", "21", "Y"),
  order.cells = F,
  subset.genes = NULL,
  euclid.dist = F,
  angle = NULL,
  x.angle = NULL,
  y.angle = 0,
  axis.rel = 1,
  base.size = 12,
  axis.title.size = 12,
  axis.text.size = 11,
  base.col = "#073642",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  text.size = 12,
  y.hide = NULL,
  tile.size = 0.1,
  tile.col = NULL,
  legend.position = "right",
  legend.height = 2,
  legend.width = 0.6,
  legend.rel = 0.9,
  legend.colour = "black",
  legend.breaks = NULL,
  legend.labels = NULL,
  legend.justification = "top",
  legend.title.position = "bottom",
  legend.title.angle = NULL,
  legend.title.rel = 0.9
)
}
\arguments{
\item{cna}{matrix of CNA values (genes by cells).}

\item{limits}{colour range. Cells >= upper limit will be the same colour, and likewise for cells <= lower limit. Default: c(-1, 1)}

\item{ratio}{aspect ratio of the panel. Default: NULL}

\item{cols}{character vector of colours to use. Default: heatCols}

\item{x.name}{x axis label. Default: 'Chromosome'}

\item{y.name}{y axis label. Default: 'Cell'}

\item{legend.title}{title of legend. Default: 'Inferred CNA `[log2 ratio`]'}

\item{x.hide}{x axis labels to hide from plot. Default: c("13", "18", "21", "Y")}

\item{order.cells}{a boolean or a list of groups if ordering should be performed within each only; this is useful if plotting multiple samples in the same panel. Cells are ordered by hierarchical clusrering. Default: F}

\item{angle}{angle of axes tick labels. x.angle and y.angle inherit from angle. Default: NULL}

\item{x.angle}{angle of x axis tick labels. If left, will inherit from angle. Default: NULL}

\item{y.angle}{angle of y axis tick labels. If left, will inherit from angle. Default: 0}

\item{axis.rel}{relative size of axes labels. Default: 1}

\item{base.size}{base text size. Default: 12}

\item{axis.title.size}{axes titles text size. Default: 12}

\item{axis.text.size}{axes labels text size. Default: 11}

\item{base.col}{base text colour. Default: '#073642'}

\item{title}{title of plot. Default: NULL}

\item{subtitle}{subtitle of plot. Default: NULL}

\item{caption}{caption of plot. Default: NULL}

\item{text.size}{base text size. Default: 12}

\item{y.hide}{y axis labels to hide from plot. Default: NULL}

\item{tile.size}{size of tile borders. No border if is equal to NULL. Default: 0.1}

\item{tile.col}{colour of tile borders. No border if is equal to NULL. Default: NULL}

\item{legend.position}{legend position on plot. Default: 'right'}

\item{legend.height}{legend height. Default: 2}

\item{legend.width}{legend width. Default: 0.6}

\item{legend.rel}{relative size of legend text. Default: 0.9}

\item{legend.colour}{legend text colour. Default: 'black'}

\item{legend.breaks}{breaks to subset.genes on colour key. Default: NULL}

\item{legend.labels}{specify labels on colour key. Default: NULL}

\item{legend.justification}{legend justification on plot. Default: 'top'}

\item{legend.title.position}{position of legend title on plot. Default: 'bottom'}

\item{legend.title.angle}{angle of legend title. Default: NULL}

\item{legend.title.rel}{relative size of legend title text. Default: 0.9}
}
\value{
a list containing $p, the ggplot plot, and $data, the data (in dataframe format).
}
\description{
Plot a heatmap of CNA values.
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[ggpubr]{rotate_axis_text}}
}
