% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaplot.R
\name{ggheatmap}
\alias{ggheatmap}
\alias{cnaplot}
\title{Plot A Heatmap with ggplot2}
\usage{
ggheatmap(
  dat,
  x = Var1,
  y = Var2,
  fill = value,
  x.num = TRUE,
  y.num = TRUE,
  limits = c(-1, 1),
  limits.find = FALSE,
  limits.symmetric = FALSE,
  cols = NULL,
  na.value = NULL,
  raster.labels = FALSE,
  raster.labels.col = "black",
  x.breaks = waiver(),
  y.breaks = waiver(),
  x.labels = waiver(),
  y.labels = waiver(),
  labels.col = "black",
  x.labels.angle = 0,
  y.labels.angle = 0,
  x.expand = c(0, 0),
  y.expand = c(0, 0),
  x.position = "bottom",
  y.position = "left",
  x.title = NULL,
  y.title = NULL,
  legend.breaks = NULL,
  legend.labels = NULL,
  legend.labels.col = "black",
  legend.width = unit(0.3, "cm"),
  legend.height = unit(0.8, "cm"),
  legend.margin = ggplot2::margin(t = -0.02, unit = "npc"),
  legend.title = "Scaled Expression",
  legend.position = "bottom",
  legend.justification = "right",
  title = waiver(),
  subtitle = waiver(),
  caption = waiver(),
  text.size = 12,
  plot.margin = ggplot2::margin(0.15, 0.15, 0.15, 0.15, "cm")
)

cnaplot(
  m,
  reorder = FALSE,
  reorder.by = NULL,
  groups = NULL,
  interorder = TRUE,
  genome = "hg19",
  dist.method = "euclidean",
  cluster.method = "complete",
  hide = c("21", "Y"),
  limits = c(-1, 1),
  y.angle = 90,
  legend.title = NULL,
  title = "Copy-number aberrations",
  axis.text.size = 12,
  legend.height = grid::unit(0.3, "cm"),
  legend.width = grid::unit(0.5, "cm"),
  cols = NULL,
  ...
)

cnaplot(
  m,
  reorder = FALSE,
  reorder.by = NULL,
  groups = NULL,
  interorder = TRUE,
  genome = "hg19",
  dist.method = "euclidean",
  cluster.method = "complete",
  hide = c("21", "Y"),
  limits = c(-1, 1),
  y.angle = 90,
  legend.title = NULL,
  title = "Copy-number aberrations",
  axis.text.size = 12,
  legend.height = grid::unit(0.3, "cm"),
  legend.width = grid::unit(0.5, "cm"),
  cols = NULL,
  ...
)
}
\arguments{
\item{dat}{A \code{data.frame} or a \code{tibble} to plot ggplot.}

\item{x}{The column name of variable mapping to x axis.}

\item{y}{The column name of variable mapping to y axis.}

\item{fill}{The column name of variable mapping to heatmap fill.}

\item{x.num}{description}

\item{y.num}{description}

\item{limits}{for the colour key; replaces out of bounds values with the nearest limit. Default: c(-1, 1)}

\item{limits.find}{If used, auto find the value limits to plot heatmap. Default: FALSE}

\item{limits.symmetric}{Used with \code{limits.find}. If used, set the value limits to be symmetrical. Default: FALSE}

\item{cols}{custom colour palette (character vector), Default: NULL}

\item{na.value}{The colot of NA value.}

\item{raster.labels}{Whether label the value in heatmap. Default: FALSE}

\item{raster.labels.col}{The color of the label in heatmap}

\item{x.breaks, y.breaks}{same as \code{breaks} in \code{\link[ggplot2]{scale_x_continuous}} or \code{\link[ggplot2]{scale_y_continuous}}}

\item{x.labels, y.labels}{same as \code{labels} in \code{\link[ggplot2]{scale_x_continuous}} or \code{\link[ggplot2]{scale_y_continuous}}}

\item{labels.col}{The color of x axis and y axis label}

\item{x.labels.angle, y.labels.angle}{The Angle of x axis or y axis label (in \eqn{[0, 360]})}

\item{x.expand, y.expand}{same as \code{expand} in \code{\link[ggplot2]{scale_x_continuous}} or \code{\link[ggplot2]{scale_y_continuous}}}

\item{x.position, y.position}{same as \code{position} in \code{\link[ggplot2]{scale_x_continuous}} or \code{\link[ggplot2]{scale_y_continuous}}}

\item{x.title, y.title}{The title of x axis or y axis}

\item{legend.breaks}{same as \code{breaks} in \code{\link[ggplot2]{scale_fill_gradientn}}}

\item{legend.labels}{same as \code{labels} in \code{\link[ggplot2]{scale_fill_gradientn}}}

\item{legend.labels.col}{The legend label color}

\item{legend.width}{legend bar width, Default: 0.5}

\item{legend.height}{legend bar height, Default: 0.3}

\item{legend.margin}{same as \code{legend.margin} in \code{\link[ggplot2]{theme}}}

\item{legend.title}{legend title, Default: NULL}

\item{legend.position}{same as \code{legend.position} in \code{\link[ggplot2]{theme}}}

\item{legend.justification}{same as \code{legend.justification} in \code{\link[ggplot2]{theme}}}

\item{title}{plot title, Default: 'Copy-number aberrations'}

\item{subtitle}{same as \code{subtitle} in \code{\link[ggplot2]{labs}}}

\item{caption}{same as \code{caption} in \code{\link[ggplot2]{labs}}}

\item{text.size}{The base size of text in the plot}

\item{plot.margin}{same as \code{plot.margin} in \code{\link[ggplot2]{theme}}}

\item{m}{CNA matrix (genes by cells). <m> can be generated using `Infercna::Infercna`.}

\item{reorder}{reorder cells using hierarchical clustering, Default: FALSE}

\item{reorder.by}{reorder cells by genes on select chromosomes or chromosome arms, e.g., c("7", "10", "1p", "19q"). Default: NULL}

\item{groups}{groups of cells to delineate and label on the plot (e.g., cell types, samples, or subclones). If <reorder> is TRUE, ordering will be performed within groups, Default: NULL}

\item{interorder}{reorder by hierarchical clustering betweeen groups, Default: TRUE}

\item{genome}{set genome to use ('hg19' or 'hg38'), Default: 'hg19'}

\item{dist.method}{distance metric for reordering, Default: 'euclidean'}

\item{cluster.method}{linkage method for reordering, Default: 'complete'}

\item{hide}{hide x-axis labels for specific chromosomes (e.g., small chromosomes whose labels would overlap flanking labels), Default: c("21", "Y")}

\item{y.angle}{y-axis labels' angle, Default: 90}

\item{axis.text.size}{x and y axes label size, Default: 12}

\item{...}{other arguments to pass to `ggheatmap`.}

\item{cna}{CNA matrix (genes by cells). It can be generated using \code{\link[Infercna]{Infercna}}.}
}
\value{
return a ggplot2 object

return a ggplot object

a `ggplot` object
}
\description{
`ggheatmap()` create a heatmap ggplot2 object.

Uses `ggplot::geom_raster` to generate a plot of copy number aberration (CNA) values. Delineates chromosomes and chromosome arms, and provides options for clustering cells or groups of cells (e.g., cell types, samples, or subclones) prior to plotting.

Uses `ggplot::geom_raster` to generate a plot of copy number aberration (CNA) values. Delineates chromosomes and chromosome arms, and provides options for clustering cells or groups of cells (e.g., cell types, samples, or subclones) prior to plotting.
}
\examples{
\dontrun{
m <- Infercna(mgh125, refCells = refCells)
malCells <- list(Malignant = setdiff(colnames(mgh125), unlist(refCells)))
groups <- c(malCells, refCells)
p <- cnaplot(m, reorder = T, groups = refCells)
}
}
