% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaScatterPlot.R
\name{cnaScatterPlot}
\alias{cnaScatterPlot}
\title{FUNCTION_TITLE}
\usage{
cnaScatterPlot(cna, cor.method = "pearson", threshold = NULL,
  cor.threshold = threshold, signal.threshold = 0.9, group = NULL,
  group.col = "magenta", hline = NULL, vline = NULL,
  bySample = FALSE, samples = NULL, sep = "-|_",
  excludeFromAvg = NULL)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}

\item{cor.method}{character string indicating the method to use for the pairwise correlations. E.g. 'pearson', 'spearman'. Default: 'pearson'}

\item{threshold}{PARAM_DESCRIPTION, Default: NULL}

\item{cor.threshold}{PARAM_DESCRIPTION, Default: threshold}

\item{signal.threshold}{PARAM_DESCRIPTION, Default: 0.9}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{group.col}{PARAM_DESCRIPTION, Default: 'magenta'}

\item{hline}{PARAM_DESCRIPTION, Default: NULL}

\item{vline}{PARAM_DESCRIPTION, Default: NULL}

\item{bySample}{PARAM_DESCRIPTION, Default: FALSE}

\item{samples}{a character vector of sample names list of character vectors (cell/column names). Can be provided if the cna matrix contains multiple samples of cells, i.e. multiple tumours, such that the cell-group correlations are calcualted for each group/tumour in turn. Default: FALSE}

\item{sep}{if bySample is TRUE and samples are NULL, split cell IDs by <sep> and take the first substring to be the sample name. Default: '-|_'}

\item{excludeFromAvg}{a character vector of cell IDs to exclude from the average tumour profile. Default: NULL}

\item{samples}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
a base R plot
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
