% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaScatterPlot.R
\name{cnaScatterPlot}
\alias{cnaScatterPlot}
\title{Visualise Malignant and Non-Malignant Subsets}
\usage{
cnaScatterPlot(
  cna,
  cor.method = "pearson",
  threshold = NULL,
  cor.threshold = threshold,
  signal.threshold = threshold,
  group = NULL,
  group.col = "magenta",
  cex = 0.5,
  hline = NULL,
  vline = NULL,
  bySample = FALSE,
  samples = NULL,
  sep = "-|_",
  excludeFromAvg = NULL,
  ...
)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}

\item{cor.method}{character string indicating the method to use for the pairwise correlations. E.g. 'pearson', 'spearman'. Default: 'pearson'}

\item{threshold}{Default: NULL}

\item{cor.threshold}{PARAM_DESCRIPTION, Default: threshold}

\item{signal.threshold}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{group.col}{PARAM_DESCRIPTION, Default: 'magenta'}

\item{hline}{PARAM_DESCRIPTION, Default: NULL}

\item{vline}{PARAM_DESCRIPTION, Default: NULL}

\item{bySample}{PARAM_DESCRIPTION, Default: FALSE}

\item{samples}{PARAM_DESCRIPTION, Default: NULL}

\item{sep}{if bySample is TRUE and samples are NULL, split cell IDs by <sep> and take the first substring to be the sample name. Default: '-|_'}

\item{excludeFromAvg}{a character vector of cell IDs to exclude from the average tumour profile. Default: NULL}

\item{...}{other arguments passed to base plot}
}
\value{
a base R plot
}
\description{
Visualise Malignant and Non-Malignant Subsets of cells. This is achieved by plotting, for each cell, its CNA signal over its CNA correlation. Please see `infercna::cnaSignal` and `infercna::cnaCor` for details.
}
