% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaScatterPlot.R
\name{cnaScatterPlot}
\alias{cnaScatterPlot}
\title{Visualise Malignant and Non-Malignant Subsets}
\usage{
cnaScatterPlot(
  cna,
  cor.method = "pearson",
  gene.quantile = NULL,
  refCells = NULL,
  samples = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}

\item{cor.method}{character string indicating the method to use for the pairwise correlations. E.g. 'pearson', 'spearman'. Default: 'pearson'}

\item{gene.quantile}{calculate CNA measures including only top / "hotspot" genes according to their squared CNA values across all cells. Value between 0 and 1 denoting the quantile of genes to include. Default: NULL}

\item{refCells}{a character vector of cell ids to exclude from average CNA profile that each cell is correlated to. You can pass reference normal cell ids to this argument if these are known. Default: NULL}

\item{samples}{if CNA correlations should be calculated within cell subgroups, provide i) a list of cell id groups, ii) a character vector of sample names to groups cells by, iii) TRUE to extract sample names from cell ids and subsequently groups. Default: NULL}

\item{verbose}{print progress messages. Default: FALSE}
}
\value{
a base R plot. If return value is saved to a variable, instead returns data points for cna correlations and cna signal in list form.
}
\description{
Visualise Malignant and Non-Malignant Subsets of cells. This is achieved by plotting, for each cell, its CNA signal over its CNA correlation. Please see `infercna::cnaSignal` and `infercna::cnaCor` for details.
}
