% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaSignal.R
\name{cnaSignal}
\alias{cnaSignal}
\title{Calculate the Means of Squared CNA Values}
\usage{
cnaSignal(cna, threshold = NULL)
}
\arguments{
\item{cna}{a matrix of gene rows by cell columns containing CNA values.}
}
\value{
a numeric vector of CNA signal values or the Mean of Squared CNA values
}
\description{
Calculate the Mean of Squared CNA Values
}
