% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{colCenter}
\alias{colCenter}
\title{FUNCTION_TITLE}
\usage{
colCenter(m)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
