% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{expandToClones}
\alias{expandToClones}
\title{FUNCTION_TITLE}
\usage{
expandToClones(modes, greaterThan = NULL)
}
\arguments{
\item{modes}{PARAM_DESCRIPTION}

\item{greaterThan}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{setNames}}
}
