% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{fetchModes}
\alias{fetchModes}
\title{FUNCTION_TITLE}
\usage{
fetchModes(
  m,
  prob = 0.95,
  coverage = 0.8,
  size = 10,
  by = "chr",
  bySampling = FALSE,
  nsamp = 2000,
  minGenes = 50,
  ...
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{prob}{PARAM_DESCRIPTION, Default: 0.95}

\item{coverage}{PARAM_DESCRIPTION, Default: 0.8vto}

\item{size}{PARAM_DESCRIPTION, Default: 10}

\item{by}{PARAM_DESCRIPTION, Default: 'chr'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
