% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{fetchModes}
\alias{fetchModes}
\title{Find small clone}
\usage{
fetchModes(
  cna,
  prob = 0.95,
  coverage = 0.8,
  size = 10,
  by = "chr",
  minGenes = 50,
  bySampling = FALSE,
  nsamp = 2000,
  ...
)
}
\arguments{
\item{cna}{a matrix of genes X cells (variables X observations) containing CNA values.}

\item{prob}{a numeric value >= 0 and <= 1; the minimum posterior probability required for an observation to be assigned to a mode. Default: 0.95}

\item{coverage}{the fraction of observations that must have a posterior probability higher than <prob> to one of two modes in order for the distribution to qualify as bimodal. Default: 0.8}

\item{size}{the minimum number of observations required to define a mode. Default: 10}

\item{by}{found clones by different gene position, Default: \code{chr}}

\item{minGenes}{The min number of genes to found clones.}

\item{bySampling, nsamp, ...}{see \code{\link[Infercna]{fitBimodal}}}
}
\value{
return a list of all small clone
}
\description{
Find all small clones by gene position
}
