% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBimodal.R
\name{fitBimodal}
\alias{fitBimodal}
\title{Fit a Bimodal Gaussian Distribution}
\usage{
fitBimodal(x, prob = 0.95, coverage = 0.8, size = 10,
  assign = FALSE, boolean = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a named numeric vector of cells/observations or a matrix of genes X cells (variables X observations). If the latter, the column means are first computed.}

\item{prob}{a numeric value >= 0 and <= 1; the minimum posterior probability required for an observation to be assigned to a mode. Default: 0.95}

\item{coverage}{the fraction of observations that must have a posterior probability higher than <prob> to one of two modes in order for the distribution to qualify as bimodal. Default: 0.8}

\item{size}{the minimum number of observations that must be assigned to a mode in order for the distribution to qualify as bimodal. Default: 10}

\item{boolean}{if set to TRUE, returns a boolean value indicating whether the distribution is bimodal. Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Fit a bimodal gaussian distribution to a set of observations.
}
\details{
DETAILS
}
\examples{
 m = useData()
 cnamat = infercna(m = m, dipcells = dipcells)
 cnamats = splitGenes(cnamat, by = 'chr')
 result = sapply(cnamats, modes, simplify = F)
 result.bool = sapply(cnamats, modes, boolean = T, simplify = F)
}
\seealso{
\code{\link[mixtools]{normalmixEM}}
}
