% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderGene.R
\name{orderGenes}
\alias{orderGenes}
\title{Order Genes by their Genomic Positions}
\usage{
orderGenes(x)
}
\arguments{
\item{x}{a matrix with gene names as rows or a character vector of gene names.}
}
\value{
ordered matrix or character vector
}
\description{
Order Genes by their Genomic Positions
}
\examples{
\dontrun{
m <- Infercna::useData()
orderGenes(m) \%>\%
  rownames() \%>\%
  head()
orderGenes(rownames(m)) \%>\% head()
}
}
