% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-infercna.R
\name{runMean}
\alias{runMean}
\title{Rolling Means}
\usage{
runMean(m, k = 100, endrule = "mean", align = "center", verbose = T)
}
\arguments{
\item{m}{a numeric vector or matrix. If the latter, each column will be processed separately.}

\item{k}{width of rolling window. Default: 100}

\item{endrule}{character string indicating how the values at the beginning and the end of the data should be treated. One of "mean", "trim", "keep", "constant". See caTools::runmean for more details. Default: 'mean'}

\item{align}{specifies whether result should be centered (default), left-aligned or right-aligned. See caTools::runmean for more details. Default: 'center'}

\item{verbose}{print progress messages. Default: TRUE}
}
\value{
a numeric vector or matrix of the same size as <m>. Only in case of endrule=trim, the output vectors will be shorter and output matrices will have fewer rows.
}
\description{
Apply a rolling window mean to a matrix or vector.
}
\seealso{
\code{\link[caTools]{runmean}}
}
