% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scalop.R
\name{theme_scalop}
\alias{theme_scalop}
\title{scalop ggtheme}
\usage{
theme_scalop(
  size = 16,
  legend.title.size = 13,
  legend.text.size = 12,
  legend.position = "right",
  legend.justification = "top"
)
}
\arguments{
\item{size}{base text size. Default: 16}

\item{legend.title.size}{legend title text size. Default: 13}

\item{legend.text.size}{legend text size. Default: 12}

\item{legend.position}{legend position. Default: 'right'}

\item{legend.justification}{legend justification. Default: 'top'}
}
\value{
function to be added to ggplot2 call
}
\description{
theme for ggplot2
}
\seealso{
\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[grid]{unit}}
}
