% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useData.R
\name{useData}
\alias{useData}
\title{FUNCTION_TITLE}
\usage{
useData(x = "125", random = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: '125'}

\item{random}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scrabble]{rowCenter}}
}
