% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useData.R
\name{useData}
\alias{useData}
\title{Load Example Data}
\usage{
useData(x = NULL, rowcenter = F)
}
\arguments{
\item{rowcenter}{center the expression matrix row-wise. Default: F}
}
\value{
centered expression matrix of genes X cells
}
\description{
Load example dataset of human single-cell RNA seq data. Cells and genes are columns and rows, respectively, and all passed QC. Matrix is row-centered.
}
